requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-chroma
langchain-ollama
ollama
chromadb
pydantic-settings
python-dotenv
unstructured[pdf,xlsx]
python-jose[cryptography]
passlib[bcrypt]
tenacity
tiktoken
lxml
pandas
openpyxl
